code/requirements.txt
numpy
pandas
Pillow
scikit_learn
scipy
torch
torchvision
tqdm
tensorboard
openpyxl